Create Download-macOS-Installer
Uses the download URL from the Download Full Installer app from https://github.com/scriptingosx/DownloadFullInstaller/releases

Using Jamf Pro, parameter 4 is where we specify the download URL obtained from Download Full Installer. Parameter 5 is where we specify the version number we are downloading. Parameter 5 is really just for being able to echo the version number we are installing for logging purposes and to give the install package a name that reflects the version of macOS we want to install.