Assign Mac to a site In Jamf Pro
Use this script to assign a Mac to a site in Jamf Pro. In Jamf Pro, use parameter 4 for the site ID. Use parameter 5 for the site name. The script will use the Mac's serial number to query Jamf Pro for the computer's Jamf Pro ID and then use the ID to assign the Mac to the site.