Update macOS_Update_Prompt_with_Swift_Dialog-Public.sh
Uses Swift Dialog to alert users that A macOS Update is required. If Swift Dialog is not installed, it will be installed.

For launch daemon and deferral PLIST replace "CompanyName" with your own organization name.

Jamf Pro Parameters: 
4 - macOS required version  5 - Release Notes  6 - Timer (in seconds)  7 - Deadline date seconds (epoch time)
8 - Required macOS major version  9 - Required macOS minor version  10 - URL or file path for Dialog icon

Release notes: https://support.apple.com/en-us/121011

For deadline date epoch time value run: 
/bin/date -jf "%Y-%m-%d %H:%M:%S" "2024-07-21 11:00:00" +%s  <--change the year-month-day hour-minute-seconds to the desired values

A meeting check is run before the user is prompted. The script checks for Microsoft Teams or Zoom in a meeting. If either app
has a meeting in progress, a 1-hour deferral is automatically set and the user is not prompted.

A macOS version check is run. If the Mac is already running a macOS version equal to or greater than the required version, the script
will exit.

The script checks for macOS installers and deletes them. For major upgrades, macOS will use the installer app instead of running the
upgrade through Software Update. The installer app needs to be deleted since non-admin users cannot run the macOS install app.

Users are allowed 3 deferrals. A deadline date is enforced.
Dialog will switch to aggressive mode after 3 deferrals or the deadline date has passed. If users select the option to update but do not
actually update a launch daemon and script are installed to prompt the user to update. The launch daemon runs every 10 minutes.

Use the "macOS Version Check - Policy file clean up" script to remove all files created by this script after successful update.

VERSION 2