Refactor Post-ZTP script for improved checks and installs
I use this script after zero touch provisioning with the first policy that runs at check-in. The script checks that all the apps that should have installed during ZTP did install and installs any that are missing. Later, it installs the remaining apps that we want installed and the company's fonts. This script calls up many policies that are in my Jamf Pro server. They can be replaced with policy custom triggers from your own server. This script also triggers the deployment of the configuration profile for platform single sign-on. Customize this script to your needs. I posted it here to be a good reference for Apple admins needing a script to verify the success of their ZTP process. The script works with a PLIST that was stored during the ZTP policy to report on failed and successful installs.